@@ -1,10 +1,14 @@
 {
-    "name": "esto-as/vero-php",
-    "description": "Library for tracking users and event with Vero's communications platform",
-    "require": {
-        "php": ">=5.3.0"
-    },
-    "autoload": {
-        "classmap": ["vero.php"]
-    }
+  "name": "esto-as/vero-php",
+  "description": "Library for tracking users and event with Vero's communications platform",
+  "require": {
+    "php": ">=7.1.0",
+    "ext-json": "*",
+    "ext-curl": "*"
+  },
+  "autoload": {
+    "classmap": [
+      "vero.php"
+    ]
+  }
 }
